% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon2-package.r
\docType{package}
\name{argon2-package}
\alias{argon2-package}
\title{argon2}
\description{
TODO
}
\author{
Drew Schmidt
}
\references{
Project URL: \url{https://github.com/wrathematics/argon2}
}
\keyword{package}

