% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon2_hash.r
\name{argon2_hash}
\alias{argon2_hash}
\title{Argon2 - Password Hashing}
\usage{
argon2_hash(
  password,
  nonce = NULL,
  type = "id",
  iterations = 1,
  memory = 2048,
  threads = 2,
  len = 64,
  as_raw = TRUE
)
}
\arguments{
\item{password}{The plaintext password to be encoded (maps to parameter `P` on RFC 9106).}

\item{nonce}{The salt to be used for the encoding (maps to parameter `S` on RFC 9106).
If not provided a random nonce of 16 bytes will be generated. This is the size
recommended by RFC 9106 for password hashing.}

\item{type}{Choice of algorithm; currently the supported choices are "i", "d" and "id".
Defaults to "id".}

\item{iterations}{A time cost. Can be any integer from 1 to 2^31 - 1. Maps to parameter `t` on
RFC 9106. Defaults to 1.}

\item{memory}{A memory cost, given in MiB. Can be any integer from 1 to 2^22 - 1. Maps to
parameter `m` on RFC 9106. Defaults to 2 GiB.}

\item{threads}{Number of threads. This affects the speed of hashing, so more is better.
Maps to parameter `p`on RFC 9106. Defaults to 2.}

\item{len}{Length of the desired output hash. Defaults to 68 bytes (512 bits).}

\item{as_raw}{TRUE (default) sets the output to a raw vector. FALSE sets output to string.}
}
\value{
An object of type `argon2.raw` consisting of a list of 2 elements: raw_hash and salt.
}
\description{
Password hashing

This uses the argon2 (i, d or id variety) hash algorithm with tweekable parameters.

The function uses by default the recommended defaults set on chapter 7.4 of RFC 9106 ("Argon2id
variant with t=1 and 2 GiB memory (...) is secure against side-channel attacks and maximizes
adversarial costs on dedicated brute-force hardware.").
See references for details and implementation source code (also bundled with this package).
}
\examples{
library(argon2)

pass <- "myPassw0rd!"
hash <- argon2_hash(pass)
hash # store this

}
