% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon2_kdf.R
\name{argon2_kdf}
\alias{argon2_kdf}
\title{Argon2 - Key Derivation function}
\usage{
argon2_kdf(password, nonce = 32L)
}
\arguments{
\item{password}{The plaintext password to derive a key from.}

\item{nonce}{The nonce to be used on the calculation. Defaults to a 32-byte random nonce.}
}
\value{
An object of type `argon2.key` consisting of a list of 2 elements: key and salt.
}
\description{
Key derivation from password

This is a wrapper for the argon2_hash function with very robust (aka, paranoid!) parameters.
}
\examples{
argon2_kdf("this is a password to derive a key from")
}
