% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon2_encode.r
\name{argon2_verify}
\alias{argon2_verify}
\title{Argon2 - Password Encoding Validation}
\usage{
argon2_verify(encoded_hash, password)
}
\arguments{
\item{encoded_hash}{A string or 'argon2.encoded' object containing an encoded password.}

\item{password}{A string representing the password to be tested.}
}
\value{
TRUE if the password matches, and FALSE otherwise.
}
\description{
validation of an encoded password given the encoded string and a password to be tested.
}
\examples{

# Checking 'argon2.encode' object
encoded <- argon2_encode("password", memory = 4)
argon2_verify(encoded, "password")
argon2_verify(encoded, "not_the_password")

# Checking a string
argon2_verify("$argon2id$v=19$m=16384,t=2,p=4$c21hbGxzYWx0$VMLGGh5t7cTbShJgAJP4gw",
              "password")
argon2_verify("$argon2id$v=19$m=16384,t=2,p=4$c21hbGxzYWx0$VMLGGh5t7cTbShJgAJP4gw",
              "not_the_password")

}
