% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_as_raw.R
\name{char_as_raw}
\alias{char_as_raw}
\title{char_as_raw}
\usage{
char_as_raw(string)
}
\arguments{
\item{string}{A string of hex encoded values.}
}
\value{
A raw vector.
}
\description{
Convert an hexadecimal encoded string to raw vector. This is different from R's
\code{charToRaw()}.  See examples for details.
}
\examples{
library(argon2)
str <- "736F6D652074657874"

charToRaw(str)

char_as_raw(str)

}
