% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_nonce.r
\name{gen_nonce}
\alias{gen_nonce}
\title{Generate a nonce}
\usage{
gen_nonce(length = 64)
}
\arguments{
\item{length}{The number of elements to return.}
}
\value{
A random raw vector.
}
\description{
Generates a random raw (unsigned char*) vector.
}
