% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_to_char.r
\name{raw_to_char}
\alias{raw_to_char}
\title{raw_to_char}
\usage{
raw_to_char(raw, upper = TRUE, spaces = FALSE)
}
\arguments{
\item{raw}{A raw vector.}

\item{upper}{Should hex digits A-F be given in uppercase?}

\item{spaces}{Should the str use spaces?}
}
\value{
A character string.
}
\description{
Convert the literal bytes of a raw (unsigned char*) to a
string representation.  This is different from R's
\code{rawToChar()}.  See examples for details.
}
\examples{
\dontrun{
library(argon2)
str <- "some text"
raw <- charToRaw(str)
raw

rawToChar(raw)
raw_to_char(raw)
}

}
