% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon2-package.r
\docType{package}
\name{argon2-package}
\alias{argon2-package}
\title{argon2}
\description{
Utilities for secure password hashing via the argon2 algorithm.
It is a relatively new hashing algorithm and is believed to be very secure.
The 'argon2' implementation included in the package is a source distribution
of the official release from the 'argon2' algorithm developers.  The package
also includes some utilities that should be useful for digest
authentication, including a wrapper of 'blake2b'.  For similar R packages,
see sodium and 'bcrypt'.
}
\references{
Project URL: \url{https://github.com/wrathematics/argon2}
}
\author{
Drew Schmidt
}
\keyword{package}
