% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argon2-package.r
\docType{package}
\name{argon2-package}
\alias{argon2-package}
\title{argon2}
\description{
Utilities for secure password hashing via the argon2 algorithm.  At the time
of writing, it is a fairly new hashing algorithm and is believed to be very
secure.  The package also includes some utilities that should be useful for
digest authentication, including a wrapper of blake2b.  For similar R
packages, see sodium and bcrypt.
}
\author{
Drew Schmidt
}
\references{
Project URL: \url{https://github.com/wrathematics/argon2}
}
\keyword{package}

