% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blake2b.r
\name{blake2b}
\alias{blake2b}
\title{blake2b}
\usage{
blake2b(x, key = NULL)
}
\arguments{
\item{key}{An optional key.}

\item{str}{Input string.}
}
\value{
The hash of the string as a raw vector.
}
\description{
A 512-bit blake2b hash implementation.
}

