% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blake2b.r
\name{blake2b}
\alias{blake2b}
\title{blake2b}
\usage{
blake2b(x, key = NULL)
}
\arguments{
\item{x}{Input to be hashed. Can be a single string or a raw vector.}

\item{key}{An optional key. Should be \code{NULL} (for no key), a single string, or a
raw vector.}
}
\value{
The hash of the string as a raw vector.
}
\description{
A 512-bit blake2b hash implementation.
}
\examples{
\dontrun{





}

}
\references{
Aumasson, J.P., Neves, S., Wilcox-O’Hearn, Z. and Winnerlein, C., 2013, June.
BLAKE2: simpler, smaller, fast as MD5. In International Conference on Applied
Cryptography and Network Security (pp. 119-135). Springer Berlin Heidelberg.
}
