% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_nonce.r
\name{gen_nonce}
\alias{gen_nonce}
\title{Generate a nonce}
\usage{
gen_nonce(length = 128)
}
\arguments{
\item{length}{The number of items to return. Default is 128 items, which
is 512 hex bits.}
}
\value{
A random raw vector.
}
\description{
Generates a random raw (unsigned char*) vector.
}

